refactor: extract envelope conversion helpers into pure module
